refactor(test): move conversion tests to integration tests
