+++ src/axis_aligned/convert.rs
@@ -262,98 +262,3 @@ where
         axis.into_signed_axis().ok_or(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::*;
-
-    const PA0_SIGNED: PositiveAxis0<u8> = PositiveAxis0::<u8>::new(0, 0, 1).unwrap();
-    const NA0_SIGNED: NegativeAxis0<u8> = NegativeAxis0::<u8>::new(0, 2, 0).unwrap();
-    const PA1_SIGNED: PositiveAxis1<u8> = PositiveAxis1::<u8>::new(0, 0, 3).unwrap();
-    const NA1_SIGNED: NegativeAxis1<u8> = NegativeAxis1::<u8>::new(0, 4, 0).unwrap();
-
-    const PA0_AXIS: Axis0<u8> = Axis::Positive(PA0_SIGNED);
-    const NA0_AXIS: Axis0<u8> = Axis::Negative(NA0_SIGNED);
-    const PA1_AXIS: Axis1<u8> = Axis::Positive(PA1_SIGNED);
-    const NA1_AXIS: Axis1<u8> = Axis::Negative(NA1_SIGNED);
-
-    const PA0_ANY: AnyAxis<u8> = AnyAxis::PositiveAxis0(PA0_SIGNED);
-    const NA0_ANY: AnyAxis<u8> = AnyAxis::NegativeAxis0(NA0_SIGNED);
-    const PA1_ANY: AnyAxis<u8> = AnyAxis::PositiveAxis1(PA1_SIGNED);
-    const NA1_ANY: AnyAxis<u8> = AnyAxis::NegativeAxis1(NA1_SIGNED);
-
-    #[test]
-    fn signed_axis_into_axis() {
-        assert_eq!(PA0_SIGNED.into_axis(), PA0_AXIS);
-        assert_eq!(PA1_SIGNED.into_axis(), PA1_AXIS);
-        assert_eq!(NA0_SIGNED.into_axis(), NA0_AXIS);
-        assert_eq!(NA1_SIGNED.into_axis(), NA1_AXIS);
-    }
-
-    #[test]
-    fn signed_axis_into_any_axis() {
-        assert_eq!(PA0_SIGNED.into_any_axis(), PA0_ANY);
-        assert_eq!(PA1_SIGNED.into_any_axis(), PA1_ANY);
-        assert_eq!(NA0_SIGNED.into_any_axis(), NA0_ANY);
-        assert_eq!(NA1_SIGNED.into_any_axis(), NA1_ANY);
-    }
-
-    #[test]
-    fn axis_into_signed_axis() {
-        assert_eq!(PA0_AXIS.into_signed_axis::<false>(), Some(PA0_SIGNED));
-        assert_eq!(NA0_AXIS.into_signed_axis::<true>(), Some(NA0_SIGNED));
-        assert_eq!(PA1_AXIS.into_signed_axis::<false>(), Some(PA1_SIGNED));
-        assert_eq!(NA1_AXIS.into_signed_axis::<true>(), Some(NA1_SIGNED));
-
-        assert_eq!(PA0_AXIS.into_signed_axis::<true>(), None);
-        assert_eq!(NA0_AXIS.into_signed_axis::<false>(), None);
-        assert_eq!(PA1_AXIS.into_signed_axis::<true>(), None);
-        assert_eq!(NA1_AXIS.into_signed_axis::<false>(), None);
-    }
-
-    #[test]
-    fn axis_into_any_axis() {
-        assert_eq!(PA0_AXIS.into_any_axis(), PA0_ANY);
-        assert_eq!(NA0_AXIS.into_any_axis(), NA0_ANY);
-        assert_eq!(PA1_AXIS.into_any_axis(), PA1_ANY);
-        assert_eq!(NA1_AXIS.into_any_axis(), NA1_ANY);
-    }
-
-    #[test]
-    fn any_axis_into_signed_axis() {
-        assert_eq!(PA0_ANY.into_signed_axis::<false, false>(), Some(PA0_SIGNED));
-        assert_eq!(NA0_ANY.into_signed_axis::<true, false>(), Some(NA0_SIGNED));
-        assert_eq!(PA1_ANY.into_signed_axis::<false, true>(), Some(PA1_SIGNED));
-        assert_eq!(NA1_ANY.into_signed_axis::<true, true>(), Some(NA1_SIGNED));
-
-        assert_eq!(PA0_ANY.into_signed_axis::<true, false>(), None);
-        assert_eq!(NA0_ANY.into_signed_axis::<false, false>(), None);
-        assert_eq!(PA1_ANY.into_signed_axis::<true, true>(), None);
-        assert_eq!(NA1_ANY.into_signed_axis::<false, true>(), None);
-    }
-
-    #[test]
-    fn any_axis_into_axis() {
-        assert_eq!(PA0_ANY.into_axis::<false>(), Some(PA0_AXIS));
-        assert_eq!(NA0_ANY.into_axis::<false>(), Some(NA0_AXIS));
-        assert_eq!(PA1_ANY.into_axis::<true>(), Some(PA1_AXIS));
-        assert_eq!(NA1_ANY.into_axis::<true>(), Some(NA1_AXIS));
-
-        assert_eq!(PA0_ANY.into_axis::<true>(), None);
-        assert_eq!(NA0_ANY.into_axis::<true>(), None);
-        assert_eq!(PA1_ANY.into_axis::<false>(), None);
-        assert_eq!(NA1_ANY.into_axis::<false>(), None);
-    }
-
-    #[test]
-    fn try_from_impls() {
-        assert_eq!(PositiveAxis0::<u8>::try_from(PA0_AXIS), Ok(PA0_SIGNED));
-        assert_eq!(NegativeAxis0::<u8>::try_from(PA0_AXIS), Err(()));
-
-        assert_eq!(Axis0::<u8>::try_from(PA0_ANY), Ok(PA0_AXIS));
-        assert_eq!(Axis1::<u8>::try_from(PA0_ANY), Err(()));
-
-        assert_eq!(PositiveAxis0::<u8>::try_from(PA0_ANY), Ok(PA0_SIGNED));
-        assert_eq!(NegativeAxis0::<u8>::try_from(PA0_ANY), Err(()));
-    }
-}
